refactor: isolated util file
